Provides a script for SSH tunneling
TL;DR
-----

Starts this repo with the current iteration of my SSH tunneling
Python script

Details
-------

Splits a script I've used for SSH tunneling out from another
repo it was a part of so I can use it and evolve it on it's own.
The code provides a script `tunnel`, written in Python, that
can establish and teardown SSH tunnels.